requirements.txt
flask>=2.3.3
numpy>=1.23.5
opencv-python>=4.8.1.78
scikit-learn>=1.2.2
scikit-image>=0.21.0
torch
torchvision
pillow>=10.0.1
joblib>=1.3.2
